+++ src/Menu.py
@@ -1,4 +1,6 @@
 from util.XML_handler import *
+import commands
+
 
 print("Welcome to VSSM v0.01")  # Removed some zeros : )
 print("type help for seeing available commands")
@@ -8,9 +10,12 @@ print("For now this program maybe helpful for storing names of data.")
 running = True
 
 while running:
-	command_input = input("VSSM > ").strip().lower().split()
-	execute(command_input)
-	if command_input[0] == "exit":
-		verify = input("Are you sure you want to leave without saving changes (Y|N) :")
-		if verify[0].strip().lower() == 'y':
-			running = False
+	try:
+		command_input = input("VSSM > ").strip().lower().split()
+		if command_input[0] == "exit":
+			verify = input("Are you sure you want to leave without saving changes (Y|N) :")
+			if verify[0].strip().lower() == 'y':
+				running = False
+		commands.execute(command_input)
+	except IndexError:
+		print("Please check the syntax of the command.")
