+++ src/util/XML_handler.py
@@ -19,48 +19,6 @@ def create_id():
 	return '-'.join(l)
 
 
-def execute(func: list):
-	"""
-	:param func: This is a list containing [function, file, Name]
-	:return: depends upon the function.
-	"""
-	func_len = len(func)
-	if func_len > 0 and func[0] in commands:
-		if (func[0] in ["create", "remove"] and func_len < 3) or (func[0] in ["save", "preview"] and func_len < 2):
-			print("More information required to execute the given command")
-			print("Type help for more information")
-		elif func[0] == "create":
-			file = file_dict[func[1]]
-			file.create(func[2])
-		elif func[0] == "remove":
-			file = file_dict[func[1]]
-			file.remove(func[2])
-		elif func[0] == "save":
-			file = file_dict[func[1]]
-			file.save()
-		elif func[0] == "preview":
-			print(file_dict[func[1]])
-		elif func[0] == "help":
-			help()
-	else:
-		print("Invalid command")
-
-
-def help():
-	print("""
-	NOTE: The commands are not case sensitive
-	NOTE: For commands save, help and exit the trailing data will be ignored.
-	NOTE: ANd for create and remove commands the trailing data will be considered as a part of name.
-	Commands:
-	create <client|product> <Name>
-	remove <client|product> <Name>
-	save
-	preview <client|product>
-	help
-	exit
-	""")
-
-
 class Client:
 	root = 'client'
 
@@ -117,8 +75,6 @@ class Product(Client):
 prod_obj = Product('util\\product.xml')
 cli_obj = Client('util\\client.xml')
 formatter = xmlformatter.Formatter()
-commands = ["help", "create", "remove", "save", "preview", "exit"]
-file_dict = {"client": cli_obj, "product": prod_obj}
 
 if __name__ == '__main__':
 	print("=====YOU ARE RUNNING THIS FILE DIRECTLY=====")
